# backend/doc/api-docs.yml
swagger: "2.0"
info:
  title: PersonalMediaVault Backend API
  description: API to access PersonalMediaVault from a web client.
  version: 2.0.0

# cSpell:ignore allof, anyof, noneof, gyan
schemes:
  - http
  - https
securityDefinitions:
  SessionTokenHeader:
    type: apiKey
    in: header
    name: x-session-token
...
paths:
  "/api/auth/login":
    post:
      summary: Login API
      description: Creates an active session from a set of credentials (user + password)
      tags:
        - auth
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: credentials
          schema:
            properties:
              username:
                type: string
                description: Media vault username (can't be blank, max 255 chars)
                example: admin
              password:
                type: string
                description: Media vault password (can't be blank, max 255 chars)
                example: password
              tfaCode:
                type: string
                description: Two factor authentication code
                example: "123456"
              duration:
                type: string
                description: "Duration of the session. Note the session can be closed before this or the vault server can restart, clearing all sessions."
                enum:
                  - day
                  - week
                  - month
                  - year

      responses:
        400:
          description: Bad request (if the password given is blank, for example)
        403:
          description: Login request denied
          schema:
            properties:
              code:
                type: string
                description: >
                  Error code:
                    * INVALID_CREDENTIALS - Invalid user or password
                    * TFA_REQUIRED - Two factor authentication is required to login, but not provided (this won't activate cooldown)
                    * INVALID_TFA_CODE - Invalid TFA code provided
                    * COOLDOWN - When you fail a login request, there is a cooldown of 5 seconds.
                enum:
                  - INVALID_CREDENTIALS
                  - TFA_REQUIRED
                  - INVALID_TFA_CODE
                  - COOLDOWN
              message:
                type: string
                description: "Error message"
                example: "Error message"
        200:
          description: Login request accepted
          schema:
            properties:
              session_id:
                type: string
                description: "Session ID. Set a cookie named 'x-session-token' with that value. Sessions expire when you log out or after 24 hours."
              vault_fingerprint:
                type: string
                description: "Vault fingerprint"
  "/api/auth/logout":
    post:
      summary: Logout API
      description: Deletes the current active session
      tags:
        - auth
      consumes:
        - application/json
      produces:
        - application/json
      parameters: []
      security:
        - SessionTokenHeader: []
      responses:
        401:
          description: Unauthorized (No valid session)
        200:
          description: Success
  "/api/account":
    get:
      summary: Get account context
      description: Gets the account context for the current session
      tags:
        - account
      consumes:
        - application/json
      produces:
        - application/json
      parameters: []
      security:
        - SessionTokenHeader: []
      responses:
        401:
          description: Unauthorized (No valid session)
        200:
          description: Username
          schema:
            properties:
              username:
                type: string
                description: Username
                example: "admin"
              title:
                type: string
                description: "Custom title for the vault"
              css:
                type: string
                description: "Custom style for the vault"
              version:
                type: string
                description: "Server version"
              root:
                type: boolean
                description: "True if the user has root access to the vault"
              write:
                type: boolean
                description: "True if the user has write access to the vault"
  "/api/account/username":
    post:
      summary: Change username
      description: Changes the username for the vault credentials
      tags:
        - account
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: credentials
          schema:
            properties:
              username:
                type: string
                description: Media vault NEW username (can't be blank, max 255 chars)
                example: admin
              password:
                type: string
                description: Media vault password (can't be blank, max 255 chars)
                example: password
        - in: header
          name: "x-auth-confirmation-tfa"
          type: string
          description: "Two factor authentication code (if applicable) to confirm the operation"
      security:
        - SessionTokenHeader: []
      responses:
        400:
          description: Bad request.
          schema:
            properties:
              code:
                type: string
                description: >
                  Error code:
                    * USERNAME_INVALID - Invalid username
                    * USERNAME_IN_USE - There is already another user with the same username
                enum:
                  - USERNAME_INVALID
                  - USERNAME_IN_USE
              message:
                type: string
                description: "Error message"
                example: "Error message"
        401:
          description: Unauthorized (No valid session)
        403:
          description: Access denied
          schema:
            properties:
              code:
                type: string
                description: >
                  Error code:
                    * AUTH_CONFIRMATION_REQUIRED_TFA - Required auth confirmation (two factor authentication)
                    * INVALID_TFA_CODE - Invalid two factor authentication code
                    * INVALID_PASSWORD - Invalid password
                    * COOLDOWN - When you fail a confirmation, there is a cooldown of 5 seconds.
                    * ACCESS_DENIED - Access denied to the API
                enum:
                  - AUTH_CONFIRMATION_REQUIRED_TFA
                  - INVALID_TFA_CODE
                  - INVALID_PASSWORD
                  - COOLDOWN
                  - ACCESS_DENIED
              message:
                type: string
                description: "Error message"
                example: "Error message"
        200:
          description: Success   
  "/api/account/password":
    post:
      summary: Change password
      description: Changes the password for the vault credentials
      tags:
        - account
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: credentials
          schema:
            properties:
              old_password:
                type: string
                description: Media vault password (can't be blank, max 255 chars)
                example: password-old
              password:
                type: string
                description: Media vault NEW password (can't be blank, max 255 chars)
                example: password
        - in: header
          name: "x-auth-confirmation-tfa"
          type: string
          description: "Two factor authentication code (if applicable) to confirm the operation"
      security:
        - SessionTokenHeader: []
      responses:
        400:
          description: Bad request. (Invalid new password)
        401:
          description: Unauthorized (No valid session)
        403:
          description: Access denied
          schema:
            properties:
              code:
                type: string
                description: >
                  Error code:
                    * AUTH_CONFIRMATION_REQUIRED_TFA - Required auth confirmation (two factor authentication)
                    * INVALID_TFA_CODE - Invalid two factor authentication code
                    * INVALID_PASSWORD - Invalid password
                    * COOLDOWN - When you fail a confirmation, there is a cooldown of 5 seconds.
                enum:
                  - AUTH_CONFIRMATION_REQUIRED_TFA
                  - INVALID_TFA_CODE
                  - INVALID_PASSWORD
                  - COOLDOWN
              message:
                type: string
                description: "Error message"
                example: "Error message"
        200:
          description: Success
  "/api/account/security":
    get:
      summary: Gets security settings
      description: Gets security settings of the account
      tags:
        - account
      consumes:
        - application/json
      produces:
        - application/json
      parameters: []
      security:
        - SessionTokenHeader: []
      responses:
        401:
          description: Unauthorized (No valid session)
        403:
          description: Access denied
        200:
          description: Success
          schema:
            properties:
              tfa:
                type: boolean
                description: "True if two factor authentication is enabled"
              tfaMethod:
                type: string
                description: >
                  Two factor authentication method.
                  Has a prefix, separated by a colon, and a list of arguments separated by colons
                  
                  For TOTP. There are 3 arguments:
                    - The HMAC algorithm (sha1, sha256, sha512)
                    - The period (seconds)
                    - A flag (0, or 1) indicating if clock skew is allowed by 1 period.
                example: "totp:sha1:60:1"
              authConfirmation:
                type: boolean
                description: "True if auth confirmation is enabled"
              authConfirmationMethod:
                type: string
                description: "Authentication confirmation preferred method. Can be 'tfa' for two factor authentication, or 'pw' for password"
                enum:
                  - tfa
                  - pw
              authConfirmationPeriodSeconds:
                type: number
                format: int32
                description: "Period to remember the auth confirmation for (in seconds)"
                example: 120
    post:
      summary: Sets security settings
      description: Sets security settings of the account
      tags:
        - account
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: settings
          schema:
            properties:
              authConfirmation:
                type: boolean
                description: "True if auth confirmation is enabled"
              authConfirmationMethod:
                type: string
                description: "Authentication confirmation preferred method. Can be 'tfa' for two factor authentication, or 'pw' for password"
                enum:
                  - tfa
                  - pw
              authConfirmationPeriodSeconds:
                type: number
                format: int32
                description: "Period to remember the auth confirmation for (in seconds)"
                example: 120
        - in: header
          name: "x-auth-confirmation-tfa"
          type: string
          description: "Two factor authentication code (if applicable) to confirm the operation"
        - in: header
          name: "x-auth-confirmation-pw"
          type: string
          description: "Account password (if applicable) to confirm the operation"
      security:
        - SessionTokenHeader: []
      responses:
        400:
          description: Bad request (invalid settings)
        401:
          description: Unauthorized (No valid session)
        403:
          description: Access denied
          schema:
            properties:
              code:
                type: string
                description: >
                  Error code:
                    * AUTH_CONFIRMATION_REQUIRED_TFA - Required auth confirmation (two factor authentication)
                    * INVALID_TFA_CODE - Invalid two factor authentication code
                    * AUTH_CONFIRMATION_REQUIRED_PW - Required auth confirmation (password)
                    * INVALID_PASSWORD - Invalid password
                    * COOLDOWN - When you fail a confirmation, there is a cooldown of 5 seconds.
                    * ACCESS_DENIED - Access denied to the API
                enum:
                  - AUTH_CONFIRMATION_REQUIRED_TFA
                  - INVALID_TFA_CODE
                  - AUTH_CONFIRMATION_REQUIRED_PW
                  - INVALID_PASSWORD
                  - COOLDOWN
                  - ACCESS_DENIED
              message:
                type: string
                description: "Error message"
                example: "Error message"
        200:
          description: Success
  "/api/account/security/tfa/totp":
    get:
      summary: Gets settings for TOTP
      description: Gets the settings to configure two factor authentication with TOTP
      tags:
        - account
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          name: issuer
          type: string
          description: "Issuer to be set for the TOTP URL"
        - in: query
          name: account
          type: string
          description: "Account to be set for the TOTP URL"
        - in: query
          name: algorithm
          type: string
          description: "HMAC algorithm"
          enum:
            - sha1
            - sha256
            - sha512
        - in: query
          name: period
          description: "TOTP period in seconds"
          type: string
          enum:
            - '30'
            - '60'
            - '120'
        - in: query
          name: skew
          type: string
          description: "Allow or disallow clock skew by 1 period"
          enum:
            - allow
            - disallow
      security:
        - SessionTokenHeader: []
      responses:
        400:
          description: Bad request.
          schema:
            properties:
              code:
                type: string
                description: >
                  Error code:
                    * INVALID_ALGORITHM - Invalid HMAC algorithm
                    * INVALID_PERIOD - Invalid TOTP period
                enum:
                  - INVALID_ALGORITHM
                  - INVALID_PERIOD
              message:
                type: string
                description: "Error message"
                example: "Error message"
        401:
          description: Unauthorized (No valid session)
        403:
          description: Access denied
        200:
          description: Success
          schema:
            properties:
              secret:
                type: string
                description: "TOTP secret"
              method:
                type: string
                description: "TOTP method"
                example: "totp:sha1:60:1"
              url:
                type: string
                description: "TOTP URL"
                example: "otpauth://totp/Example:alice@google.com?secret=JBSWY3DPEHPK3PXP&issuer=Example"
              qr:
                type: string
                description: "QR code as a base-64 image URL"
                example: "data:image/png;base64,"
    post:
      summary: Enables TOTP
      description: Enables two factor authentication with TOTP
      tags:
        - account
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: settings
          schema:
            properties:
              secret:
                type: string
                description: "TOTP secret"
              method:
                type: string
                description: "TOTP method"
                example: "totp:sha1:60:1"
              password:
                type: string
                description: "Account password, required to confirm the operation"
              code:
                type: string
                description: "TOTP code to confirm the operation"
                example: "123456"
      security:
        - SessionTokenHeader: []
      responses:
        400:
          description: Bad request.
          schema:
            properties:
              code:
                type: string
                description: >
                  Error code:
                    * TFA_ENABLED - Two factor authentication is already enabled
                    * INVALID_TOTP_SECRET - Invalid TOTP secret
                    * INVALID_TOTP_OPTIONS - Invalid TOTP options in the method string
                    * INVALID_TOTP_CODE - Invalid TOTP confirmation code
                enum:
                  - TFA_ENABLED
                  - INVALID_TOTP_SECRET
                  - INVALID_TOTP_OPTIONS
                  - INVALID_TOTP_CODE
              message:
                type: string
                description: "Error message"
                example: "Error message"
        401:
          description: Unauthorized (No valid session)
        403:
          description: Access denied
          schema:
            properties:
              code:
                type: string
                description: >
                  Error code:
                    * ACCESS_DENIED - Access denied to the API
                    * INVALID_PASSWORD - Invalid password
                enum:
                  - ACCESS_DENIED
                  - INVALID_PASSWORD
              message:
                type: string
                description: "Error message"
                example: "Error message"
        200:
          description: Success
...
  "/api/tags":
    get:
      summary: Get tags
      description: Get full tag list 
      tags:
        - tags
      consumes:
        - application/json
      produces:
        - application/json
      parameters: []
      security:
        - SessionTokenHeader: []
      responses:
        401:
          description: Unauthorized (No valid session)
        200:
          description: Tag list
          schema:
            type: array
            items:
              properties:
                id:
                  type: number
                  format: int64
                  description: Tag ID
                name:
                  type: string
                  description: Tag name
...
  "/api/tasks":
    get:
      summary: Get active tasks
      description: Get active tasks
      tags:
        - tasks
      consumes:
        - application/json
      produces:
        - application/json
      parameters: []
      security:
        - SessionTokenHeader: []
      responses:
        401:
          description: Unauthorized (No valid session)
        200:
          description: Task list
          schema:
            type: array
            items:
              $ref: '#/definitions/TaskInfo'
...